Updated  functions to read sosi files and get right points
Updated lesSosiful so that it doesnt collect points outside limit and
that when it reads height curvs, it uses same height for alle the
points.
Updated Plan class to take inn variables from filklasse.
Filled hoydedata.txt with all points and added minX, minY, maxX, minX
values on the top.